# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TJ-T-Mart API",
        "description": "REST API for TJ-T-Mart — Dormitory Mini-Mart System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/galon": {
            "get": {
                "tags": [
                    "Galon"
                ],
                "summary": "List user's galon transactions",
                "operationId": "1769e85c14430002f39efc062fb24098",
                "responses": {
                    "200": {
                        "description": "Paginated galon history",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Galon"
                ],
                "summary": "Create galon water order",
                "description": "Prices calculated server-side. Galon Baru + Isi = 45000, Isi Ulang = 18000. Delivery fee 3000/gallon for 'antar'.",
                "operationId": "94cb51d4ade93cc89d28f56a199e57a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GalonRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Galon order created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/galon": {
            "get": {
                "tags": [
                    "Admin",
                    "Galon"
                ],
                "summary": "Admin: list all galon transactions",
                "operationId": "3fe943092eed3e45fb4a59da55737683",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "metode_pengiriman",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ambil",
                                "antar"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin galon list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/galon/{id}/status": {
            "put": {
                "tags": [
                    "Admin",
                    "Galon"
                ],
                "summary": "Admin: update galon order status",
                "operationId": "a110505cbbaaad25522f0cba95f44042",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "paid",
                                            "delivering",
                                            "completed",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/mart": {
            "get": {
                "tags": [
                    "Mart"
                ],
                "summary": "List active marts",
                "description": "Returns all active marts with lokasi_delivery and produk counts.",
                "operationId": "75e35731ea6a9956e6d08467bcb1a765",
                "responses": {
                    "200": {
                        "description": "Mart list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mart/{id}": {
            "get": {
                "tags": [
                    "Mart"
                ],
                "summary": "Get mart detail",
                "operationId": "b54eaf60395507886a6ce9c194509531",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mart detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mart not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifikasi": {
            "get": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "List user notifications",
                "description": "Includes unread_count in meta. Paginated 20/page.",
                "operationId": "783b107f8f9b93855a7c86fc1f5f6d99",
                "responses": {
                    "200": {
                        "description": "Notification list with meta.unread_count",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifikasi/{id}/read": {
            "put": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "Mark notification as read",
                "description": "IDOR-protected: only the notification owner can mark it read.",
                "operationId": "b3e37bfe6bd852a80ef3e858f73f5941",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marked as read",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notification not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifikasi/read-all": {
            "put": {
                "tags": [
                    "Notifikasi"
                ],
                "summary": "Mark all notifications as read",
                "operationId": "2763de394c365cd61b724a41418dd214",
                "responses": {
                    "200": {
                        "description": "All marked as read",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiSuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "OK"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "GalonRequest": {
                "required": [
                    "nama_galon",
                    "jumlah",
                    "metode_pembayaran",
                    "metode_pengiriman"
                ],
                "properties": {
                    "nama_galon": {
                        "type": "string",
                        "example": "Galon 19L (Isi Ulang)",
                        "enum": [
                            "Galon Baru + Isi",
                            "Galon 19L (Isi Ulang)"
                        ]
                    },
                    "jumlah": {
                        "type": "integer",
                        "example": 2,
                        "maximum": 50,
                        "minimum": 1
                    },
                    "metode_pembayaran": {
                        "type": "string",
                        "example": "COD",
                        "enum": [
                            "COD",
                            "MIDTRANS"
                        ]
                    },
                    "metode_pengiriman": {
                        "type": "string",
                        "example": "antar",
                        "enum": [
                            "ambil",
                            "antar"
                        ]
                    },
                    "catatan": {
                        "type": "string",
                        "example": "Antar ke kamar A-101"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}